refactor: convert example scripts to typescript
