+++ package.json
@@ -141,7 +141,7 @@
     "ts-jest": "^29.1.0",
     "ts-loader": "^9.2.6",
     "ts-node": "10.9.1",
-    "tsconfig-paths": "^3.12.0",
+    "tsconfig-paths": "^3.14.2",
     "typescript": "^4.9.0",
     "weak-napi": "^2.0.2",
     "webpack": "^5.69.0",
