+++ packages/certificate-summary/example.ts
@@ -1,14 +1,14 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const { buildCertificateSummary } = require('../../dist/packages/certificate-summary/src/index');
-const en10168Certificate = require('../../fixtures/EN10168/v0.4.1/valid_cert.json');
-const coaCertificate = require('../../fixtures/CoA/v1.1.0/valid_cert.json');
-const eCoCCertificate = require('../../fixtures/E-CoC/v1.0.0/valid_cert.json');
+/* eslint-disable no-console */
+import coaCertificate from '../../fixtures/CoA/v1.1.0/valid_cert.json';
+import eCoCCertificate from '../../fixtures/E-CoC/v1.0.0/valid_cert.json';
+import en10168Certificate from '../../fixtures/EN10168/v0.4.1/valid_cert.json';
+import { buildCertificateSummary, CertificateSummary } from './src/index';
 
 (async function (argv) {
   try {
     const schemaType = argv[2] || 'en10168-schemas';
 
-    let summary;
+    let summary: CertificateSummary | null = null;
     if (schemaType.startsWith('en10168')) {
       summary = await buildCertificateSummary(en10168Certificate);
     } else if (schemaType.startsWith('e-coc')) {
